+++ README.md
@@ -1,0 +1,10 @@
+Благодарю за предыдущее ревью и желаю хорошего дня и прекрасного настроения! 
+test_main_logo = тест перехода по логотипам
+test_order = тест заказа по первому сценарию 
+test_order_second = тест заказа по второму сценарию 
+test_samokat = тест ответов на вопросы 
+
+page_object
+arenda_page = страница уточнения информации аренды 
+order_page = страница уточнения заказа 
+samokat_page = главная страница самоката
